requirements.txt
flask
requests
os
python-dotenv
markdown